Passenger(DAO, DAOImpl, Service, ServiceImpl) add

+++ src/main/java/com/javaschool/ev/domain/Passenger.java
@@ -9,7 +9,7 @@ import javax.persistence.*;
 public class Passenger {
 
     @Id
-    private Long passengerID;
+    private int passengerID;
     private String firstName;
     private String lastName;
     //private SimpleDateFormat birthDate;
